fix: rename requestController to uppercase, move auth dtos to a directory
