+++ src/main/java/com/github/nhatoriginal/spring/controller/AuthController.java
@@ -1,9 +1,9 @@
 package com.github.nhatoriginal.spring.controller;
 
 import com.github.nhatoriginal.spring.constant.Endpoint;
-import com.github.nhatoriginal.spring.dto.AuthLoginDto;
-import com.github.nhatoriginal.spring.dto.AuthRegisterDto;
-import com.github.nhatoriginal.spring.dto.AuthResponseDto;
+import com.github.nhatoriginal.spring.dto.auth.AuthLoginDto;
+import com.github.nhatoriginal.spring.dto.auth.AuthRegisterDto;
+import com.github.nhatoriginal.spring.dto.auth.AuthResponseDto;
 import com.github.nhatoriginal.spring.service.AuthService;
 
 import jakarta.validation.Valid;
@@ -21,18 +21,15 @@ public class AuthController {
     this.service = service;
   }
 
+  @PostMapping(Endpoint.Auth.REGISTER)
+  public ResponseEntity<String> register(
+      @Validated @RequestBody AuthRegisterDto request) {
+    return ResponseEntity.ok(service.register(request));
+  }
 
-
-    @PostMapping(Endpoint.Auth.REGISTER)
-    public ResponseEntity<String> register(
-           @Validated  @RequestBody AuthRegisterDto request
-            ) {
-        return ResponseEntity.ok(service.register(request));
-    }
-    @PostMapping(Endpoint.Auth.LOGIN)
-    public ResponseEntity<AuthResponseDto> login(
-           @Validated  @RequestBody AuthLoginDto authLoginDto
-    ) {
-        return ResponseEntity.ok(service.login(authLoginDto));
-    }
+  @PostMapping(Endpoint.Auth.LOGIN)
+  public ResponseEntity<AuthResponseDto> login(
+      @Validated @RequestBody AuthLoginDto authLoginDto) {
+    return ResponseEntity.ok(service.login(authLoginDto));
+  }
 }
