+++ src/main/java/com/github/nhatoriginal/spring/dto/auth/AuthResponseDto.java
@@ -1,4 +1,4 @@
-package com.github.nhatoriginal.spring.dto;
+package com.github.nhatoriginal.spring.dto.auth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
@@ -7,7 +7,6 @@ import lombok.Data;
 @Data
 @Builder
 public class AuthResponseDto {
-    @JsonProperty("access_token")
-    private String accessToken;
+  @JsonProperty("access_token")
+  private String accessToken;
 }
-
